requirements.txt
Flask==1.1.2
opencv-python==4.5.2.52
Pillow==8.2.0
Flask-RESTful==0.3.9
numpy==1.20.2
gunicorn==20.1.0
firebase-admin==5.0.2
